+++ src/Measurements.jl
@@ -31,10 +31,9 @@ import Base: show
 export Measurement, ±, stdscore, weightedmean
 
 # Define the "Derivatives" type, used inside "Measurement" type.  This should be
-# a lightweight and immutable dictionary.  In Julia 0.5 we just use
-# "Base.ImmutableDict", in previous versions we define the type just like
-# "Base.ImmutableDict".
-if VERSION >= v"0.5-dev+1910"
+# a lightweight and immutable dictionary.  When "ImmutableDict" is available use
+# it, otherwise define a type just like it.
+if isdefined(Base, :ImmutableDict)
     const Derivatives = Base.ImmutableDict
 else
     include("derivatives-type.jl")
@@ -101,6 +100,44 @@ function weightedmean(iterable)
     w = [inv(el.err)^2 for el in iterable]
     invsumw = inv(sum(w))
     return Measurement(dot(v, w)*invsumw, sqrt(invsumw))
+end
+
+# Derivative and Gradient
+derivative{F<:AbstractFloat, T<:AbstractFloat}(a::Measurement{F},
+                                               tag::Tuple{T, T, Float64}) =
+                                                   get(a.der, tag, zero(F))
+
+"""
+    derivative(x::Measurement, y::Measurement)
+
+Return the value of the total derivative of `x` with respect to the independent
+measurement `y`, calculated on the nominal value of `y`.  Return `0.0` if `x`
+does not depend on `y`.
+
+Use `Measurements.gradient` to calculate the gradient of `x` with respect to an
+arrays of independent measurements.
+"""
+derivative(a::Measurement, b::Measurement) =
+    derivative(a, (b.val, b.err, b.tag))
+
+"""
+    gradient(x::Measurement, [y::AbstractArray{Measurement}])
+
+Return the gradient of `x` with respect to the independent measurements
+contained in `y` array, each calculated on the nominal value of the
+corresponding measurement.  Return `0.0` for the variables on which `x` does not
+depend.
+
+Use `Measurements.derivative` to calculate the total derivative of `x` with
+respect to a single independent measurement.
+"""
+function gradient{F<:AbstractFloat, T<:Measurement}(a::Measurement{F},
+                                                    b::AbstractArray{T})
+    out = similar(b, F)
+    for i in eachindex(b)
+        out[i] = derivative(a, b[i])
+    end
+    return out
 end
 
 include("conversions.jl")